HEAD-566 Upgrade from Honeydew to Lime RC7 fails - capi cannot find the pg_dump file - using the wrong filename to locate it

+++ zones/capi/95-zone-datasets.sh
@@ -21,30 +21,35 @@ cd /opt/smartdc/$app-repo
 # Create some directories into $app-data
 mkdir -p /opt/smartdc/$app-data/log
 mkdir -p /opt/smartdc/$app-data/tmp/pids
+
 # Remove and symlink directories:
 if [[ ! -n ${KEEP_DATA_DATASET} ]]; then
   mv /opt/smartdc/$app/config /opt/smartdc/$app-data/config
 else
   rm -Rf /opt/smartdc/$app/config
 fi
+
 rm -Rf /opt/smartdc/$app/log
 rm -Rf /opt/smartdc/$app/tmp
 rm -Rf /opt/smartdc/$app/config
 ln -s /opt/smartdc/$app-data/log /opt/smartdc/$app/log
 ln -s /opt/smartdc/$app-data/tmp /opt/smartdc/$app/tmp
 ln -s /opt/smartdc/$app-data/config /opt/smartdc/$app/config
+
 # Save REVISION:
 if [[ ! -n ${KEEP_DATA_DATASET} ]]; then
   echo "${REVISION}">/opt/smartdc/$app-data/REVISION
 fi
 echo "${REVISION}">/opt/smartdc/$app/REVISION
+
 # Save VERSION (Updates based on this):
 APP_VERSION=$(/opt/local/bin/git describe --tags)
-echo "${APP_VERSION}">/opt/smartdc/$app-data/VERSION
+if [[ ! -n ${KEEP_DATA_DATASET} ]]; then
+  echo "${APP_VERSION}">/opt/smartdc/$app-data/VERSION
+fi
 echo "${APP_VERSION}">/opt/smartdc/$app/VERSION
+
 # Cleanup build products:
 cd /root/
 rm -Rf /opt/smartdc/$app-repo
 rm /root/$app-app-timestamp
-
-
